+++ miden-crypto/src/merkle/smt/mod.rs
@@ -109,12 +109,12 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     ///
     /// This also recomputes all hashes between the leaf (associated with the key) and the root,
     /// updating the root itself.
-    fn insert(&mut self, key: Self::Key, value: Self::Value) -> Self::Value {
-        let old_value = self.insert_value(key.clone(), value.clone()).unwrap_or(Self::EMPTY_VALUE);
+    fn insert(&mut self, key: Self::Key, value: Self::Value) -> Result<Self::Value, MerkleError> {
+        let old_value = self.insert_value(key.clone(), value.clone())?.unwrap_or(Self::EMPTY_VALUE);
 
         // if the old value and new value are the same, there is nothing to update
         if value == old_value {
-            return value;
+            return Ok(value);
         }
 
         let leaf = self.get_leaf(&key);
@@ -125,7 +125,7 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
 
         self.recompute_nodes_from_index_to_root(node_index, Self::hash_leaf(&leaf));
 
-        old_value
+        Ok(old_value)
     }
 
     /// Recomputes the branch nodes (including the root) from `index` all the way to the root.
@@ -166,10 +166,14 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     /// be queried with [`MutationSet::root()`]. Once a mutation set is returned,
     /// [`SparseMerkleTree::apply_mutations()`] can be called in order to commit these changes to
     /// the Merkle tree, or [`drop()`] to discard them.
+    ///
+    /// # Errors
+    /// If mutations would exceed [`MAX_LEAF_ENTRIES`] (1024 entries) in a leaf, returns
+    /// [`MerkleError::TooManyLeafEntries`].
     fn compute_mutations(
         &self,
         kv_pairs: impl IntoIterator<Item = (Self::Key, Self::Value)>,
-    ) -> MutationSet<DEPTH, Self::Key, Self::Value> {
+    ) -> Result<MutationSet<DEPTH, Self::Key, Self::Value>, MerkleError> {
         self.compute_mutations_sequential(kv_pairs)
     }
 
@@ -178,7 +182,7 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     fn compute_mutations_sequential(
         &self,
         kv_pairs: impl IntoIterator<Item = (Self::Key, Self::Value)>,
-    ) -> MutationSet<DEPTH, Self::Key, Self::Value> {
+    ) -> Result<MutationSet<DEPTH, Self::Key, Self::Value>, MerkleError> {
         use NodeMutation::*;
 
         let mut new_root = self.root();
@@ -210,10 +214,17 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
                     // none at all), as multi-leaves should be really rare.
                     let existing_leaf = acc.clone();
                     self.construct_prospective_leaf(existing_leaf, k, v)
+                        .expect("current leaf should be valid")
                 })
             };
 
-            let new_leaf = self.construct_prospective_leaf(old_leaf, &key, &value);
+            let new_leaf =
+                self.construct_prospective_leaf(old_leaf, &key, &value).map_err(|e| match e {
+                    SmtLeafError::TooManyLeafEntries { actual } => {
+                        MerkleError::TooManyLeafEntries { actual }
+                    },
+                    other => panic!("unexpected SmtLeaf::insert error: {:?}", other),
+                })?;
 
             let mut new_child_hash = Self::hash_leaf(&new_leaf);
 
@@ -257,12 +268,12 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
             new_pairs.insert(key, value);
         }
 
-        MutationSet {
+        Ok(MutationSet {
             old_root: self.root(),
             new_root,
             node_mutations,
             new_pairs,
-        }
+        })
     }
 
     /// Applies the prospective mutations computed with [`SparseMerkleTree::compute_mutations()`] to
@@ -273,6 +284,8 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     /// [`MerkleError::ConflictingRoots`] with a two-item [`Vec`]. The first item is the root hash
     /// the `mutations` were computed against, and the second item is the actual current root of
     /// this tree.
+    /// If mutations would exceed [`MAX_LEAF_ENTRIES`] (1024 entries) in a leaf, returns
+    /// [`MerkleError::TooManyLeafEntries`].
     fn apply_mutations(
         &mut self,
         mutations: MutationSet<DEPTH, Self::Key, Self::Value>,
@@ -309,7 +322,7 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
         }
 
         for (key, value) in new_pairs {
-            self.insert_value(key, value);
+            self.insert_value(key, value)?;
         }
 
         self.set_root(new_root);
@@ -370,10 +383,13 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
 
         let mut reverse_pairs = Map::new();
         for (key, value) in new_pairs {
-            if let Some(old_value) = self.insert_value(key.clone(), value) {
-                reverse_pairs.insert(key, old_value);
-            } else {
-                reverse_pairs.insert(key, Self::EMPTY_VALUE);
+            match self.insert_value(key.clone(), value)? {
+                Some(old_value) => {
+                    reverse_pairs.insert(key, old_value);
+                },
+                None => {
+                    reverse_pairs.insert(key, Self::EMPTY_VALUE);
+                },
             }
         }
 
@@ -416,7 +432,11 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     fn remove_inner_node(&mut self, index: NodeIndex) -> Option<InnerNode>;
 
     /// Inserts a leaf node, and returns the value at the key if already exists
-    fn insert_value(&mut self, key: Self::Key, value: Self::Value) -> Option<Self::Value>;
+    fn insert_value(
+        &mut self,
+        key: Self::Key,
+        value: Self::Value,
+    ) -> Result<Option<Self::Value>, MerkleError>;
 
     /// Returns the value at the specified key. Recall that by definition, any key that hasn't been
     /// updated is associated with [`Self::EMPTY_VALUE`].
@@ -439,12 +459,16 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     /// Because this method is for a prospective key-value insertion into a specific leaf,
     /// `existing_leaf` must have the same leaf index as `key` (as determined by
     /// [`SparseMerkleTree::key_to_leaf_index()`]), or the result will be meaningless.
+    ///
+    /// # Errors
+    /// If inserting the key-value pair would exceed [`MAX_LEAF_ENTRIES`] (1024 entries) in a leaf,
+    /// returns [`SmtLeafError::TooManyLeafEntries`].
     fn construct_prospective_leaf(
         &self,
         existing_leaf: Self::Leaf,
         key: &Self::Key,
         value: &Self::Value,
-    ) -> Self::Leaf;
+    ) -> Result<Self::Leaf, SmtLeafError>;
 
     /// Maps a key to a leaf index
     fn key_to_leaf_index(key: &Self::Key) -> LeafIndex<DEPTH>;